+++ lib/get.dart
@@ -7,6 +7,8 @@ export 'src/bottomsheet/bottomsheet.dart';
 export 'src/snackbar/snack_route.dart';
 export 'src/state/get_state.dart';
 export 'src/root/root_widget.dart';
+export 'src/routes/default_route.dart';
+export 'src/routes/get_route.dart';
 export 'src/routes/observers/route_observer.dart';
 export 'src/routes/transitions_type.dart';
 export 'src/platform/platform.dart';